finalização do projeto 4 | Iniciando o projeto 5 JPA

+++ 04-JPA-API/src/br/com/fiap/jpa/dao/impl/BebidaDAOImpl.java
@@ -1,14 +1,13 @@
 package br.com.fiap.jpa.dao.impl;
 
-import java.util.Calendar;
-import java.util.GregorianCalendar;
-
+import javax.naming.CommunicationException;
 import javax.persistence.EntityManager;
-import javax.persistence.Id;
 
 import br.com.fiap.entity.Bebida;
-import br.com.fiap.entity.TipoBebida;
 import br.com.fiap.jpa.dao.BebidaDAO;
+import br.com.fiap.jpa.exception.BebidaNaoEncontradaException;
+import br.com.fiap.jpa.exception.CommitException;
+
 
 public class BebidaDAOImpl implements BebidaDAO{
 
@@ -22,30 +21,41 @@ public class BebidaDAOImpl implements BebidaDAO{
 	@Override
 	public Bebida cadastrar(Bebida bebida) {
 		em.persist(bebida);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 		return bebida;
 	}
 
 	@Override
 	public Bebida atualizar(Bebida bebida) {
 		em.merge(bebida);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 		return bebida;
 	}
 
 	@Override
-	public void remover(int codigo) {
+	public void remover(int codigo) throws BebidaNaoEncontradaException {
 		Bebida bebida = buscar(codigo);
+		if(bebida == null) 
+			throw new BebidaNaoEncontradaException();
+		
 		em.remove(codigo);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 	}
 
 	@Override
 	public Bebida buscar(int codigo) {
 		return em.find(Bebida.class, codigo);
 		
+	}
+
+	@Override
+	public void commit() throws CommitException {
+		try{
+			em.getTransaction().begin();
+			em.getTransaction().commit();
+		}catch(Exception e) {
+			if(em.getTransaction().isActive()) {
+			   em.getTransaction().rollback();
+			e.printStackTrace();
+			throw new CommitException("Erro ao gravar");
+			}
+		}
 	}
 }
